requirements.txt
alembic==1.4.3
blinker==1.4
click==7.1.2
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Reuploaded==0.3.2
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
markdown2==2.3.9
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
